Consolidate shared valibot schemas
