+++ source/SIL.AppBuilder.Portal/src/lib/projects/index.ts
@@ -1,6 +1,5 @@
-import { paginateSchema } from '$lib/table';
 import { isAdminForOrg, isSuperAdmin } from '$lib/utils/roles';
-import { idSchema } from '$lib/valibot';
+import { idSchema, langtagRegex, paginateSchema } from '$lib/valibot';
 import type { Session } from '@auth/sveltekit';
 import type { Prisma } from '@prisma/client';
 import * as v from 'valibot';
@@ -77,45 +76,6 @@ export const projectSearchSchema = v.object({
   page: paginateSchema,
   search: v.string()
 });
-
-//language tag regex sourced from: https://stackoverflow.com/a/60899733
-export const langtagRegex = new RegExp(
-  '^(' +
-    '(' + // grandfathered
-  /* irregular */ '(en-GB-oed|i-ami|i-bnn|i-default|i-enochian|i-hak|i-klingon|i-lux|i-mingo|i-navajo|i-pwn|i-tao|i-tay|i-tsu|sgn-BE-FR|sgn-BE-NL|sgn-CH-DE)' +
-    '|' +
-  /* regular */ '(art-lojban|cel-gaulish|no-bok|no-nyn|zh-guoyu|zh-hakka|zh-min|zh-min-nan|zh-xiang)' +
-    ')' +
-    '|' + // langtag
-    '(' +
-    '(' +
-    //language
-    ('([A-Za-z]{2,3}(-' +
-      //extlang
-      '([A-Za-z]{3}(-[A-Za-z]{3}){0,2})' +
-      ')?)|[A-Za-z]{4}|[A-Za-z]{5,8})') +
-    '(-' +
-    '([A-Za-z]{4})' +
-    ')?' + //script
-    '(-' +
-    '([A-Za-z]{2}|[0-9]{3})' +
-    ')?' + //region
-    '(-' +
-    '([A-Za-z0-9]{5,8}|[0-9][A-Za-z0-9]{3})' +
-    ')*' + //variant
-    //extension
-    '(-' +
-    '(' +
-  /* singleton */ ('[0-9A-WY-Za-wy-z]' + '(-[A-Za-z0-9]{2,8})+)') +
-    ')*' +
-    '(-' +
-    '(x(-[A-Za-z0-9]{1,8})+)' +
-    ')?' + //private use
-    ')' +
-    '|' +
-    '(x(-[A-Za-z0-9]{1,8})+)' +
-    ')$'
-);
 
 const projectSchemaBase = v.object({
   Name: v.pipe(v.string(), v.minLength(1)),
