--- source/SIL.AppBuilder.Portal/src/lib/table.ts
@@ -1,8 +1,0 @@
-import * as v from 'valibot';
-
-export const paginateSchema = v.object({
-  page: v.number(),
-  size: v.number()
-});
-
-export type PaginateSchema = typeof paginateSchema;
