+++ source/SIL.AppBuilder.Portal/src/routes/(authenticated)/users/+page.server.ts
@@ -1,7 +1,6 @@
 import { localizeHref } from '$lib/paraglide/runtime';
-import { paginateSchema } from '$lib/table';
 import { isSuperAdmin } from '$lib/utils/roles';
-import { idSchema } from '$lib/valibot';
+import { idSchema, paginateSchema } from '$lib/valibot';
 import type { Prisma } from '@prisma/client';
 import { error, redirect } from '@sveltejs/kit';
 import { DatabaseWrites, prisma } from 'sil.appbuilder.portal.common';
